Refactoring all Dune Queryies through dune.py
