+++ tests/queries/test_slippage_accounting.py
@@ -2,11 +2,11 @@ import unittest
 
 from duneapi.types import DuneQuery, Network, QueryParameter
 
-from src.fetch.period_slippage import (
+from src.models.slippage import (
     SplitSlippages,
     slippage_query,
 )
-from src.fetch.token_list import fetch_trusted_tokens
+from src.fetch.token_list import get_trusted_tokens
 from src.models.accounting_period import AccountingPeriod
 from tests.db.pg_client import ConnectionType, DBRouter
 
@@ -35,7 +35,7 @@ class TestDuneAnalytics(unittest.TestCase):
                 QueryParameter.date_type("StartTime", period.start),
                 QueryParameter.date_type("EndTime", period.end),
                 QueryParameter.text_type("TxHash", "0x"),
-                QueryParameter.text_type("TokenList", ",".join(fetch_trusted_tokens())),
+                QueryParameter.text_type("TokenList", ",".join(get_trusted_tokens())),
             ],
             description="",
             query_id=-1,
